pyproject.toml
[tool.poetry]
name = "data-engineer-test"
version = "0.1.0"
description = ""
authors = ["Marina"]
readme = "README.md"
packages = [{include = "data_engineer_test"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pymysql = "^1.0.3"
pd = "^0.0.4"
pandas = "^2.0.2"
python-dotenv = "^1.0.0"
cryptography = "^41.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
